chore: rename story files
